Add a product: Creation of product class
This is the class that will represent the product. It has three fields
for the instance (product number, price, and quantity), one static field
for the class (number of products) to keep track of the product count,
and four properties (Name, Price, Quantity, IsNull). Additionally, it
includes a default constructor.